finished sentinel

+++ src/interfaces/IGate.cairo
@@ -1,3 +1,7 @@
+use starknet::ContractAddress;
+
+use aura::utils::wadray::Wad;
+
 #[abi]
 trait IGate {
     // getters
@@ -5,13 +9,13 @@ trait IGate {
     fn get_shrine() -> ContractAddress;
     fn get_asset() -> ContractAddress;
     // external
-    fn enter(user: ContractAddress, trove_id: u64, asset_amt: u64) -> Wad;
-    fn exit(user: ContractAddress, trove_id: u64, yang_amt: Wad) -> u64;
+    fn enter(user: ContractAddress, trove_id: u64, asset_amt: u128) -> Wad;
+    fn exit(user: ContractAddress, trove_id: u64, yang_amt: Wad) -> u128;
     fn kill();
     // view
-    fn get_total_assets() -> u64;
+    fn get_total_assets() -> u128;
     fn get_total_yang() -> Wad;
     fn get_asset_amt_per_yang() -> Wad;
-    fn preview_enter(asset_amt: u64) -> Wad;
-    fn preview_exit(yang_amt: Wad) -> u64;
+    fn preview_enter(asset_amt: u128) -> Wad;
+    fn preview_exit(yang_amt: Wad) -> u128;
 }
